+++ CHANGELOG.md
@@ -15,8 +15,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Width of Gantt chart is now calculated on browser window width, making it full screen
 - work in progress on [clorichel/ganttlab issues](https://gitlab.com/clorichel/ganttlab/issues?scope=all&state=opened&utf8=%E2%9C%93&label_name%5B%5D=Feature)...
 
+### Changed
+- Today vertical axis on the Gantt chart is now dashed for readability
+
 ### Fixed
 - A padding problem causing useless horizontal scrolling on the page
+- Today is now always on the Gantt chart, even if it is out of range or not on an existing tick
 
 ## 0.2.0 - 2016-11-05
 ### Added